requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
clipboard==0.0.4
colorama==0.4.6
decorator==4.4.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
MouseInfo==0.1.3
moviepy==1.0.3
numpy==2.3.1
pillow==11.3.0
proglog==0.1.12
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pystray==0.19.5
python-dotenv==1.1.1
pytube==15.0.0
pytubefix==9.3.0
pytweening==1.2.0
requests==2.32.4
six==1.17.0
tqdm==4.67.1
urllib3==2.5.0
